:truck: Move Context into config module
