+++ data_safe_haven/administration/users/guacamole_users.py
@@ -1,8 +1,7 @@
 import pathlib
 from collections.abc import Sequence
 
-from data_safe_haven.config import DSHPulumiConfig, SREConfig
-from data_safe_haven.context import Context
+from data_safe_haven.config import Context, DSHPulumiConfig, SREConfig
 from data_safe_haven.external import AzureApi, AzurePostgreSQLDatabase
 from data_safe_haven.infrastructure import SREProjectManager
