+++ data_safe_haven/config/__init__.py
@@ -1,3 +1,4 @@
+from .context import Context
 from .context_manager import ContextManager
 from .dsh_pulumi_config import DSHPulumiConfig
 from .dsh_pulumi_project import DSHPulumiProject
@@ -5,6 +6,7 @@ from .shm_config import SHMConfig
 from .sre_config import SREConfig
 
 __all__ = [
+    "Context",
     "ContextManager",
     "DSHPulumiConfig",
     "DSHPulumiProject",
