--- data_safe_haven/context/__init__.py
@@ -1,5 +1,0 @@
-from .context import Context
-
-__all__ = [
-    "Context",
-]
