+++ data_safe_haven/infrastructure/programs/imperative_shm.py
@@ -1,5 +1,4 @@
-from data_safe_haven.config import SHMConfig
-from data_safe_haven.context import Context
+from data_safe_haven.config import Context, SHMConfig
 from data_safe_haven.exceptions import (
     DataSafeHavenAzureError,
     DataSafeHavenMicrosoftGraphError,
